Convert statuses::MAP to uint16_t - it doesn't have to be `long` for the PID calcs anymore

+++ speeduino/load_source.h
@@ -38,6 +38,6 @@ static inline uint16_t getLoad(LoadSource algorithm, const statuses &current) {
     return fast_div32_16((uint32_t)current.MAP * 100UL, current.EMAP);
   } else {
     // LOAD_SOURCE_MAP (the default). Aka Speed Density
-    return (uint16_t)current.MAP;
+    return current.MAP;
   }
 }
