+++ speeduino/secondaryTables.cpp
@@ -24,7 +24,7 @@ static inline bool fuelModeCondSwitchRpmActive(const config10 &page10, const sta
 
 static inline bool fuelModeCondSwitchMapActive(const config10 &page10, const statuses &current) {
   return (page10.fuel2SwitchVariable == FUEL2_CONDITION_MAP)
-      && ((uint16_t)(int16_t)current.MAP > page10.fuel2SwitchValue);
+      && (current.MAP > page10.fuel2SwitchValue);
 }
 
 static inline bool fuelModeCondSwitchTpsActive(const config10 &page10, const statuses &current) {
@@ -102,7 +102,7 @@ static inline bool sparkModeCondSwitchRpmActive(const config10 &page10, const st
 
 static inline bool sparkModeCondSwitchMapActive(const config10 &page10, const statuses &current) {
   return (page10.spark2SwitchVariable == SPARK2_CONDITION_MAP)
-      && ((uint16_t)(int16_t)current.MAP > page10.spark2SwitchValue);
+      && (current.MAP > page10.spark2SwitchValue);
 }
 
 static inline bool sparkModeCondSwitchTpsActive(const config10 &page10, const statuses &current) {
